secured endpoints using getSession

+++ pages/api/users/[userid]/index.js
@@ -4,6 +4,7 @@ import User from '../../../../models/user';
 import { Record, String } from 'runtypes';
 import { NotFoundError } from "../../../../errors/notFound.error";
 import handleError from "../../../../utils/handleError";
+import { getSession } from 'next-auth/react';
 
 const getUserByIdRunType = Record({
     query: Record({
@@ -12,44 +13,53 @@ const getUserByIdRunType = Record({
 })
 
 export default async function handler(req, res) {
+    const session = await getSession({ req })
     await dbConnect();
     const { method } = req;
     switch (method) {
         case "GET":
-            try {
-                const validatedRequest = getUserByIdRunType.check(req)
-                const { userid } = validatedRequest.query
+            if (session) {
+                try {
+                    const validatedRequest = getUserByIdRunType.check(req)
+                    const { userid } = validatedRequest.query
 
-                const foundUser = await User.findOne({ _id: userid });
-                // Check if any user found
-                if (foundUser) {
-                    return res.status(200).json({
-                        success: true,
-                        User: foundUser
-                    });
-                } else {
-                    throw new NotFoundError('No user found')
+                    const foundUser = await User.findOne({ _id: userid });
+                    // Check if any user found
+                    if (foundUser) {
+                        return res.status(200).json({
+                            success: true,
+                            User: foundUser
+                        });
+                    } else {
+                        throw new NotFoundError('No user found')
+                    }
+                } catch (error) {
+                    handleError(error, res)
                 }
-            } catch (error) {
-                handleError(error, res)
+            } else {
+                console.log('Not signed in')
             }
         case "DELETE":
-            try {
-                const validatedRequest = getUserByIdRunType.check(req)
-                const { userid } = validatedRequest.query
+            if (session) {
+                try {
+                    const validatedRequest = getUserByIdRunType.check(req)
+                    const { userid } = validatedRequest.query
 
-                const foundUser = await User.findOne({ _id: userid });
+                    const foundUser = await User.findOne({ _id: userid });
 
-                if (foundUser) {
-                    await User.findByIdAndDelete({ _id: userid })
+                    if (foundUser) {
+                        await User.findByIdAndDelete({ _id: userid })
 
-                    console.log('deleted user')
-                } else {
-                    throw new NotFoundError('No user found')
+                        console.log('deleted user')
+                    } else {
+                        throw new NotFoundError('No user found')
+                    }
+
+                } catch (error) {
+                    handleError(error, res)
                 }
-
-            } catch (error) {
-                handleError(error, res)
+            } else {
+                console.log('Not signed in')
             }
 
         default:
